requirements-drone.txt
autopep8==1.3.4
docopt==0.6.2
git+https://github.com/ronaldlam/FirstPartyLibs.git@staging
matplotlib==2.2.2
pipdeptree==0.13.1
psutil==5.4.5
pylint==1.8.3
pytest-cov==2.5.1
pytest-mock==1.9.0
python-dotenv==0.8.2
rope==0.10.7
setuptools-scm==3.1.0
wheel==0.31.1
